@@ -21,8 +21,7 @@
   },
   "license": "MIT",
   "require": {
-    "php": ">=7.1.3",
-    "ext-json": "*",
-    "hail/singleton": "^1.0.0"
+    "php": ">=8.0.0",
+    "ext-json": "*"
   }
 }
